==> 023-ph-poverty-mapping/ph-poverty-mapping/notebooks/04_osm_model.py <==
#!/usr/bin/env python
# coding: utf-8

# # OpenStreetMaps (OSM) Model

# ## Imports

# In[1]:


import wandb
import logging
import warnings
import numpy as np
import pandas as pd
import data_utils
import model_utils
import sys
sys.path.insert(0, '../utils')


logger = logging.getLogger()
logger.setLevel(logging.CRITICAL)

wandb.init(project="tm-poverty-prediction")

get_ipython().run_line_magic('load_ext', 'autoreload')
get_ipython().run_line_magic('autoreload', '2')


# ## File Locations

# In[2]:


bucket_name = 'tm-geospatial'
directory = 'poverty-prediction-datasets'

# data_dir = '../data/zoom17/'
data_dir = '../data/'
ntl_summary_stats_file = data_dir+'nightlights_summary_stats.csv'
dhs_indicators_file = data_dir+'dhs_indicators.csv'

osm_roads_file = data_dir+'osm_roads.csv'
osm_buildings_file = data_dir+'osm_buildings.csv'
osm_pois_file = data_dir+'osm_pois.csv'


# ## Load Datasets

# In[3]:


# Load DHS indicators and nighttime lights
dhs_indicators = pd.read_csv(dhs_indicators_file)
ntl_summary_stats = pd.read_csv(ntl_summary_stats_file)

# Load OSM datasets
roads = pd.read_csv(osm_roads_file)
buildings = pd.read_csv(osm_buildings_file)
pois = pd.read_csv(osm_pois_file)

# Define feature columns
road_cols = roads.columns[1:]
building_cols = buildings.columns[1:]
poi_cols = pois.columns[1:]
ntl_cols = ntl_summary_stats.columns[1:]

# Get list of columns
osm_cols = list(road_cols) + list(building_cols) + list(poi_cols)
osm_ntl_cols = list(osm_cols) + list(ntl_cols)


# ## Combine Features

# In[4]:


# Merge roads + buildings + pois
osm = roads.merge(buildings, on='DHSCLUST')
osm = osm.merge(pois, on='DHSCLUST')
print("Shape of osm dataframe: {}".format(osm.shape))

# Merge osm + nightlights
osm_ntl = osm.merge(ntl_summary_stats, on='DHSCLUST')
print("Shape of osm+ntl dataframe: {}".format(osm_ntl.shape))

# merge dhs + osm
dhs = dhs_indicators.merge(
    osm_ntl, left_on='Cluster number', right_on='DHSCLUST')


# ## Spearman and Pearsons Correlation

# In[5]:


data_utils.plot_corr(
    data=dhs,
    features_cols=osm_cols,
    indicator='Wealth Index',
    max_n=50,
    figsize=(5, 13)
)


# ## Machine Learning Pipeline

# ### Configuration

# In[6]:


# Scoring metrics
scoring = {
    'r2': data_utils.pearsonr2,
    'rmse': data_utils.rmse
}

# Indicators of interest
indicators = [
    'Wealth Index',
    'Education completed (years)',
    'Access to electricity',
    'Access to water (minutes)'
]


# ## OSM Features + Nighttime Lights

# In[ ]:


predictions = model_utils.evaluate_model(
    data=dhs,
    feature_cols=osm_ntl_cols,
    indicator_cols=indicators,
    wandb=wandb,
    scoring=scoring,
    model_type='random_forest',
    refit='r2',
    search_type='random',
    n_splits=5,
    n_iter=10,
    plot_importance=False,
    verbose=2
)


# ## OSM Features Only

# In[7]:


predictions = model_utils.evaluate_model(
    data=dhs,
    feature_cols=osm_cols,
    indicator_cols=indicators,
    scoring=scoring,
    wandb=None,
    model_type='random_forest',
    refit='r2',
    search_type='random',
    n_splits=5,
    n_iter=10,
    plot_importance=True
)


# ## Road Features

# In[9]:


predictions = model_utils.evaluate_model(
    data=dhs,
    feature_cols=road_cols,
    indicator_cols=indicators,
    scoring=scoring,
    wandb=None,
    model_type='random_forest',
    refit='r2',
    search_type='random',
    n_splits=5,
    n_iter=10,
    plot_importance=True
)


# ## Buildings

# In[10]:


predictions = model_utils.evaluate_model(
    data=dhs,
    feature_cols=building_cols,
    indicator_cols=indicators,
    scoring=scoring,
    wandb=None,
    model_type='random_forest',
    refit='r2',
    search_type='random',
    n_splits=5,
    n_iter=10,
    plot_importance=True
)


# ## Points of Interest

# In[11]:


predictions = model_utils.evaluate_model(
    data=dhs,
    feature_cols=poi_cols,
    indicator_cols=indicators,
    scoring=scoring,
    wandb=None,
    model_type='random_forest',
    refit='r2',
    search_type='random',
    n_splits=5,
    n_iter=10,
    plot_importance=True
)
